.migration-backup/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
openai>=1.0.0
presidio-analyzer>=2.2.0
spacy>=3.7.0

